Rework contribution based on discussion with @protz
See discussion in https://github.com/AeneasVerif/eurydice/pull/388

This commit:

1. Reverts AstOfLlbc.ml unsigned-cast wrapping and add_assign fix.
2. Keeps Cleanup1.ml wrapping inlining extensions and signed wrapping functions in eurydice_glue.h.
3. Adds informal guarantee statement to README.md.

+++ include/eurydice_glue.h
@@ -387,7 +387,7 @@ static inline uint64_t core_num__u64__rotate_left(uint64_t x0, uint32_t x1) {
 }
 
 static inline void core_ops_arith__i32__add_assign(int32_t *x0, int32_t *x1) {
-  *x0 = (int32_t)((uint32_t)*x0 + (uint32_t)*x1);
+  *x0 = *x0 + *x1;
 }
 
 static inline uint8_t Eurydice_bitand_pv_u8(uint8_t *p, uint8_t v) {
